refactor(frontend): move landing page constants and types to dedicated files, update imports ml-59
