+++ apps/frontend/src/pages/landing/constants/about-links.constants.ts
@@ -1,0 +1,13 @@
+import { AppRoute } from "~/libs/enums/enums.js";
+
+import { type LinkGroup } from "./../types/types.js";
+
+const ABOUT_LINKS: LinkGroup = {
+	links: [
+		{ label: "Terms & Conditions", to: AppRoute.ROOT },
+		{ label: "Privacy Policy", to: AppRoute.ROOT },
+	],
+	title: "About",
+};
+
+export { ABOUT_LINKS };
