+++ apps/frontend/src/pages/landing/constants/header-links.constants.ts
@@ -1,0 +1,10 @@
+import { AppRoute } from "~/libs/enums/enums.js";
+
+import { type NavLink } from "./../types/types.js";
+
+const HEADER_LINKS: NavLink[] = [
+	{ label: "Features", to: AppRoute.ROOT },
+	{ label: "Contacts", to: AppRoute.ROOT },
+];
+
+export { HEADER_LINKS };
