+++ apps/frontend/src/pages/landing/types/link-group.type.ts
@@ -1,0 +1,8 @@
+import { type NavLink } from "./nav-link.type.js";
+
+type LinkGroup = {
+	links: NavLink[];
+	title: string;
+};
+
+export { type LinkGroup };
