+++ apps/frontend/src/pages/landing/types/types.ts
@@ -1,0 +1,2 @@
+export { type LinkGroup } from "./link-group.type.js";
+export { type NavLink } from "./nav-link.type.js";
